// src/features/news/types.tsx
export interface INoticiasNormalizadas {
    id: number;
    titulo: string;
    descripcion: string;
    fecha: number | string;
    esPremium: boolean;
    imagen: string;
    descripcionCorta?: string;
}

export interface IModal{
    modal:INoticiasNormalizadas | null,
    setModal:(modal: INoticiasNormalizadas | null) => void;
}